Function call handling.
Relatively conservative argument dependency detection:
Mark all defined "parameter" registers as used.
For now, all potential return registers are marked as defined after call
so they may be incaccurately added to the data flow graph. The reason
this is handled that way is to be certain that DFG equivalence makes
sure that the functions are equivalent (although I this is true no
matter the exact assumptions, as long they are the same between the
functions being compared)

+++ include/ppc2cpp/data_flow/PpcInstructionUD.hpp
@@ -17,13 +17,13 @@ const uint64_t supported_register_types =
   PPC_OPERAND_SPR |
   PPC_OPERAND_GQR;
 
-extern const std::unordered_set<int32_t> fixedPointInsn;
-extern const std::unordered_set<int32_t> floatInsn;
-extern const std::unordered_set<int32_t> loadInsn;
-extern const std::unordered_set<int32_t> storeInsn;
-extern const std::unordered_set<int32_t> branchInsn;
-extern std::unordered_set<int32_t> nonStoreInsn;
-extern std::unordered_set<int32_t> supportedInsn;
+extern const std::unordered_set<uint64_t> fixedPointInsn;
+extern const std::unordered_set<uint64_t> floatInsn;
+extern const std::unordered_set<uint64_t> loadInsn;
+extern const std::unordered_set<uint64_t> storeInsn;
+extern const std::unordered_set<uint64_t> branchInsn;
+extern std::unordered_set<uint64_t> nonStoreInsn;
+extern std::unordered_set<uint64_t> supportedInsn;
 
 class InstructionUD {
 public:
